requirements.txt
astroid==2.2.5
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
defusedxml==0.7.1
ecdsa==0.13.2
Flask==2.3.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==3.0.3
future==0.17.1
idna==3.4
importlib-metadata==6.6.0
isort==4.3.18
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
lazy-object-proxy==1.4.0
MarkupSafe==2.1.3
mccabe==0.6.1
oauthlib==3.2.2
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.3.1
PyJWT==2.1.0
pylint==2.3.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.12.0
social-auth-core==4.4.2
SQLAlchemy==2.0.15
typed-ast==1.4.2
typing_extensions==4.6.2
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.11.1
zipp==3.15.0
